pyproject.toml
[tool.poetry]
name = "proj-1"
version = "0.1.0"
description = ""
authors = ["Alexander Lukichev <allukichev@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
python-dotenv = "^1.0.1"
requests = "^2.32.3"
openpyxl = "^3.1.5"
pandas = "^2.2.3"
pytest = "^8.3.3"


[tool.poetry.group.lint.dependencies]
flake8 = "^7.1.1"
mypy = "^1.13.0"
black = "^24.10.0"
isort = "^5.13.2"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
pytest-cov = "^6.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
disallow_untyped_defs = true
warn_return_any = true
exclude = 'venv'


[tool.black]
line-length = 119
exclude = '\.git'

[tool.isort]
line_length = 119

